Fix bug in usage metrics that caused a negative count to occur
When multiple service instances are registered to ONLY a single node,
removing that node would cause the "service-names" usage entry to be
decremented for each service instance instead of only once. This is
because the trigger for decrementing the service name is if no other
service instances with that name remain. Since deleting a node causes
all services registered to be deleted in the same transaction, and we
only write usage metrics at Commit() time, each individual deletion
passed that check.

To fix, we keep track of what service names have already been seen, and
only allow the usage to be updated once.